% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_code.R
\name{compile_code}
\alias{compile_code}
\title{Builds the starr executable}
\usage{
compile_code()
}
\value{
No output if build is successful, it raises errors otherwise
}
\description{
\code{compile_code()} will execute \code{cmake} and \code{make} commands to be able to execute
the C++ code in the \code{src} directory.
}
