% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_code.R
\name{run_code}
\alias{run_code}
\title{Run starr MCMC sampler}
\usage{
run_code(
  data,
  response,
  covariates,
  IDStations,
  x,
  y,
  Time,
  nu = 0.5,
  phi_prior = c(2, 4),
  n_iter = 5000,
  burn_in = 1000
)
}
\arguments{
\item{data}{a data frame that contains spatial and temporal information, with columns corresponding to a response variable, covariate variables, 2D geographical information (such as latitude, longitude), and location specification (can be numerical).}

\item{response}{A string containing the column name in \code{data} corresponding to the response variable.}

\item{covariates}{A vector of strings containing the column names in \code{data} that shall be included as covariates.}

\item{IDStations}{A string containing the column name in \code{data} that specifies the location corresponding to each row.}

\item{x}{A string containing the column name in \code{data} that specifies the x-coordinate.}

\item{y}{A string containing the column name in \code{data} that specifies the y-coordinate.}

\item{Time}{A string containing the column name in \code{data} that specifies the time variable. Can be in any sortable format.}

\item{nu}{A numeric value specifying nu for the Matérn correlation function. If not set, the default value is set to 0.5.}

\item{phi_prior}{a numeric vector of size 2 specifying the prior hyperparameters to be set for phi. If not set, the default value is
set to \code{c(2, 4)}.}

\item{n_iter}{An integer value specifying the number of iterations to run. If not set, the default is 5000.}

\item{burn_in}{A numeric value specifying number of burn-in iterations. If not set, the default  is 1000.}
}
\value{
A list with the following components:
\itemize{
\item{\strong{beta_samples} returns a data matrix of size \code{n_covariates} x \code{n_iter}. The order of the rows correspond to the order specified by the user in \code{covariates}.}
\item{\strong{rho_samples} returns a data vector of size \code{n_iter}.}
\item{\strong{sigw_samples} returns a data vector of size \code{n_iter}.}
\item{\strong{sigeps_samples} returns a data vector of size \code{n_iter}.}
\item{\strong{sig0_samples} returns a data vector of size \code{n_iter}.}
\item{\strong{phi_samples} returns a data vector of size \code{n_iter}.}
\item{\strong{mu0_samples} returns a data matrix of size \code{number_locations} x \code{n_iter}. The rows are sorted by time, i.e., after sorting \code{data} by time, \code{mu0_samples[1]} will correspond to the first location
appearing at time t = 1, \code{mu0_samples[2]} to the second location at time t=1, and so on.}
\item{\strong{y_fitted} returns a data matrix of size \code{number_locations} x \code{T} of the fitted values. The rows are sorted by time, i.e., after sorting \code{data} by time, \code{y_fitted[1]} will correspond to the first location
appearing at time t = 1, \code{y_fitted[2]} to the second location at time t=1, and so on.}
}
}
\description{
This function calls the starr executable that is build through \code{compile_code()} function, and print
the results on the terminal for the parameters of the spatio-temporal AR(1) model. Additionally, it returns the
MCMC chains of the parameters.
}
